Sensor and Display working
Lots has happened.
I played around with the code.py under the display folder just to get a count to update. and I got that working. 

So now in the main code.py, The display is setup with Lux, IR and "Raw Counts" and updates/displays data read-in from the sensor!